refactor(document): services singulares + reorg de tasks/tests
